@@ -13,31 +13,34 @@
 		"format": "prettier --write ."
 	},
 	"devDependencies": {
-		"@kayord/tw-plugin": "^0.0.7",
-		"@kayord/ui": "^0.3.5",
-		"@sveltejs/adapter-auto": "^3.0.0",
+		"@kayord/ui": "^0.17.2",
+		"@sveltejs/adapter-auto": "^6.0.0",
 		"@sveltejs/adapter-static": "^3.0.1",
 		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@types/eslint": "8.56.0",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
-		"autoprefixer": "^10.4.16",
-		"eslint": "^8.56.0",
-		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"postcss": "^8.4.32",
-		"postcss-load-config": "^5.0.2",
+		"@sveltejs/vite-plugin-svelte": "^5.0.3",
+		"@tailwindcss/vite": "^4.1.3",
+		"@threlte/core": "^8.0.1",
+		"@threlte/extras": "^9.1.1",
+		"@types/three": "^0.175.0",
+		"@typescript-eslint/eslint-plugin": "^8.29.1",
+		"@typescript-eslint/parser": "^8.29.1",
+		"eslint": "^9.24.0",
+		"eslint-config-prettier": "^10.1.2",
+		"eslint-plugin-svelte": "^3.5.1",
+		"postcss-load-config": "^6.0.1",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
-		"prettier-plugin-tailwindcss": "^0.5.9",
-		"svelte": "^4.2.7",
-		"svelte-check": "^3.6.0",
-		"tailwindcss": "^3.3.6",
+		"prettier-plugin-tailwindcss": "^0.6.11",
+		"svelte": "^5.26.1",
+		"svelte-check": "^4.1.5",
+		"tailwindcss": "^4.1.3",
+		"three": "^0.175.0",
 		"tslib": "^2.4.1",
+		"tw-animate-css": "^1.2.5",
 		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
-		"vitest": "^1.2.0"
+		"vite": "^6.2.6",
+		"vitest": "^3.1.1"
 	},
-	"type": "module"
+	"type": "module",
+	"packageManager": "pnpm@10.8.0+sha512.0e82714d1b5b43c74610193cb20734897c1d00de89d0e18420aebc5977fa13d780a9cb05734624e81ebd81cc876cd464794850641c48b9544326b5622ca29971"
 }
